requirements.txt
asgiref==3.3.4
Django==3.2
django-extensions==3.1.3
djangorestframework==3.12.4
pygraphviz==1.7
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
